Add support for runtime type testing and typecasting to S2.
Adds new isa, instanceof and as operators to S2. All three accept
an object as the left operand and a class name as the right operand.

isa checks to see if $obj is an instance of the given class or any
of its subclasses. (Pronounced "is a".)

instanceof checks only for the named class, ignoring subclasses.

"as" is the typecast operator, which will do some runtime type checking
and, if successful, return the object as a new static type. If the
types are incompatible, "as" returns a null object.

This is a cut down version of the patch I submitted some time ago in
bug 1823. The changes to the scope-handling code are not included.
Also, the runtime library changes are slightly different to the
original patch.

These modifications have been tested against the regression test suite
as well as all existing LiveJournal layouts. No regressions were
apparent.