Fake S2's lexical scoping for the PHP target.
This uses a similar decoration trick to that used by the JavaScript backend. However, rather than special-casing "this" (which is incorrect, since a layer might create a local $this in a child scope), I've added a per-variable flag to scopes which causes them not to be decorated.

The system-supplied $this must remain undecorated because the codegen for the $.whatever shorthand can't "see" the function scope with which to decorate the reference.