Add a new layertype "markup".
markup sits between i18nc and layout in the layer order and is a child of core. Layout can now be a child of either core or markup, since markup is optional.

The motivation for this is to allow the separation of the code that generates the HTML from the code that generates the CSS. (or XML and XSLT, or whatever) Had I done this a few years ago, LJ's Dear Diary and Quite Lickable could have shared a common ancestor to generate the markup, and sixhtml could have been a markup layer with each of the TypePad themes hanging off it as layouts.

It's probably too late for LJ to go down this road now, so it's completely optional and LiveJournal/FotoBilder will go on working just fine ignoring this layertype. However, one possible migration path for old S2-based applications is to create a completely empty markup layer and reparent existing layouts to hang off that, thus avoiding the UI complications of having layouts hanging off both cores and markups.