"while" and "for" loops for S2.
These were originally left out of the language to avoid people accidentally making layers that loop forever. However, people have found clever ways to shoot themselves in the foot despite the restricted set of loop statements and as a result a maximum execution time was added. This means there's little reason not to allow the full set of control flow constructs present in general purpose languages, which'll make life easier for people implementing sorting algorithms and whatnot.

The syntax is the same as most C-like languages except that it is not permissible to omit any of the expressions in the for loop. Scoping is Perl-like: a vardecl in the for loop header makes the variable local to the loop.

A flag is provided on the checker to retain the old behavior (no for and while loops allowed) should a particular caller wish to retain the restrictions. In this case, the compiler emits a much more helpful error message than it did before when a loop is encountered!